Prevent the gui from crashing if `build_imported_metadata_extractor` fails
This concerns the issue:
https://github.com/CellProfiler/CellProfiler/issues/4279

This is a rather simple fix that just captures the JavaException to prevent the whole module to crash.
I think it might be better to raise a 'ValidationError' and make sure `get_metadata_keys` is run as part
of `validate_module` and only allow attempt to run `update_table` if the module is validly configured.